moves layout into components directory (it's technically shared across pages), moves all shared directories to src, replace imports with relative import
