+++ src/components/Layout/index.tsx
@@ -16,6 +16,7 @@ export const Layout: React.FC<Props> = ({
       <meta charSet="utf-8" />
       <meta name="viewport" content="initial-scale=1.0, width=device-width" />
     </Head>
+
     <header>
       <nav>
         <Link href="/">
@@ -35,7 +36,9 @@ export const Layout: React.FC<Props> = ({
         </Link>
       </nav>
     </header>
+
     {children}
+
     <footer>
       <hr />
       <span>I'm here to stay (Footer)</span>
