+++ src/components/ListDetail.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { User } from '../types';
+import { User } from '@typings/index';
 
 interface ListDetailProps {
   item: User;
