+++ src/graphql/recipe/allRecipesQuery.graphql
