+++ src/hooks/recipe/useAllRecipes.ts
@@ -1,7 +1,7 @@
 import { QueryHookResult, useQuery } from 'react-apollo-hooks';
-import { allRecipesQuery } from '../../utils/graphqlLoader';
-import { AllRecipesQuery } from '../../types/generated';
-import { Recipe } from '../../types/generated';
+import { allRecipesQuery } from '@utils/graphqlLoader';
+import { AllRecipesQuery } from '@typings/generated';
+import { Recipe } from '@typings/generated';
 
 type AllRecipesQueryResult = Omit<QueryHookResult<AllRecipesQuery, {}>, 'data'>;
 type UseAllRecipes = () => AllRecipesQueryResult & {
