+++ src/typings/index.ts
