+++ src/typings/shared.ts
